AD-299: Partial payments for Giftcards
